streamlit/requirements.txt
streamlit
tensorflow
pandas
Pillow
numpy